app/requirements.txt
Django==2.2.*
djangorestframework==3.9.*
flake8==3.6.*
psycopg2==2.7.*
whitenoise==5.0.1
requests